requirements.txt
akshare
pandas
mplfinance
matplotlib
plotly
pyyaml
